Fix codeclimate issues
