+++ webview/src/experiments/components/table/header/TableHeaderCellContents.tsx
@@ -1,8 +1,9 @@
-import React, { createRef, useEffect, useRef } from 'react'
+import React, { useEffect } from 'react'
 import cx from 'classnames'
 import { ColumnType, Experiment } from 'dvc/src/experiments/webview/contract'
 import { flexRender, Header } from '@tanstack/react-table'
 import { SortOrder } from './ContextMenuContent'
+import { ColumnResizer, ResizerHeight } from './ColumnResizer'
 import styles from '../styles.module.scss'
 import {
   Draggable,
@@ -86,7 +87,7 @@ export const TableHeaderCellContents: React.FC<{
   onDragLeave: DragFunction
   canResize: boolean
   setMenuSuppressed: (menuSuppressed: boolean) => void
-  resizerHeight: { normal: string; hover: string }
+  resizerHeight: ResizerHeight
 }> = ({
   header,
   sortEnabled,
@@ -103,8 +104,6 @@ export const TableHeaderCellContents: React.FC<{
   setMenuSuppressed,
   resizerHeight
 }) => {
-  const resizer = createRef<HTMLDivElement>()
-  const hoveringResizer = useRef(false)
   const isTimestamp = header.headerGroup.id === ColumnType.TIMESTAMP
   const columnIsResizing = header.column.getIsResizing()
 
@@ -115,21 +114,6 @@ export const TableHeaderCellContents: React.FC<{
       document.body.classList.remove(styles.isColumnResizing)
     }
   }, [columnIsResizing])
-
-  const resetResizer = () => {
-    if (resizer.current) {
-      if (!columnIsResizing) {
-        resizer.current.style.height = resizerHeight.normal
-        return
-      }
-      if (hoveringResizer.current) {
-        return
-      }
-      setTimeout(() => {
-        resetResizer()
-      }, 300)
-    }
-  }
 
   return (
     <>
@@ -154,26 +138,11 @@ export const TableHeaderCellContents: React.FC<{
         onDragLeave={onDragLeave}
       />
       {canResize && (
-        <div
-          {...{ onMouseDown: header.getResizeHandler() }}
-          ref={resizer}
-          onMouseEnter={() => {
-            setMenuSuppressed(true)
-            hoveringResizer.current = true
-            if (resizer.current) {
-              resizer.current.style.height = resizerHeight.hover
-            }
-          }}
-          onMouseLeave={() => {
-            setMenuSuppressed(false)
-            resetResizer()
-            hoveringResizer.current = false
-          }}
-          className={cx(
-            styles.columnResizer,
-            columnIsResizing && styles.isResizing
-          )}
-          role="separator"
+        <ColumnResizer
+          setMenuSuppressed={setMenuSuppressed}
+          resizerHeight={resizerHeight}
+          header={header}
+          columnIsResizing={columnIsResizing}
         />
       )}
     </>
